Further warning/error cleanup. Rename Codable
This was a run of the Swift 4.2 migrator on the `ModuloKit` target.
Resulted in force unwrap cleanup (inserted from the initial `modulo`
target Swift 4.2 migration) as well as renaming `Codable` and `
Decodable` to `ELCodable` and `ELDecodable` in addition to associated
errors.

Since Apple took over/stole the `Codable`/`Decodable` names, `ELCodable`
now sacrafices the good names for namespaced ones.