Work on supporting configuration in Modulo file
Testing is still required for
- migrating to `options` key in `modulo` file for old versions where
that did not exist
- setting & reading defaults is a command that behaves similarly to the
other commands

Changes include parsing the spec before any commands are run or
initialized so that if we _do_ have a spec our options can be loaded &
applied to all commands. This is primarily required for `verbose` since
it impacts all the commands.