Improve `defaults` command
Rename `verboseOutput` to `alwaysVerbose`. It seemed like ELCLI was
swallowing commands that had matching prefixes so `--verbose` and
`--verboseOutput`` were being captured by the same block when that was
not the desired result.

Additionally added unit tests for the defaults command around
migrations, passing trash values, and its use being applied as expected.