Update for feedback & remove `--branch` option
Make for the `--unmanaged` flag on `add` to have an optional argument of
a commit or a branch. However this lead to complications around other
flags being passed such as `add blah --unmanaged -u -v`.

To resolve that issue, the properties on `Option` were move to `public`
so the add command could try its best to still apply arguments when
they were inadvertantly swallowed by the now `--unmanaged` optional
value.

Also the `unmanagedValue` is tracked in the `.modulo` file like `branch`
was but is more generic supporting an entire commit hash or a branch
name